requirements.txt
# Requirements for Linux Learning Telegram Bot
python-telegram-bot==20.7
httpx==0.25.2
supabase==2.3.0
python-dotenv==1.0.0
Pillow==10.1.0
PyPDF2==3.0.1
